packages/allocator/pyproject.toml
[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[project]
authors = [
  {name = "Elizabeth Berrigan", email = "eberrigan@salk.edu"},
  {name = "Talmo Pereira", email = "talmo@salk.edu"},
  {name = "Andrew Park", email = "hep003@ucsd.edu"},
]
dependencies = [
  "flask>=2.2",
  "werkzeug>=2.2",
  "flask_sqlalchemy>=2.5",
  "sqlalchemy>=1.3",
  "psycopg2-binary>=2.9",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "requests>=2.28",
  "Flask-HTTPAuth>=4.7",
  "boto3>=1.26",
]
description = "VM Allocator Service Package for Lablink"
name = "lablink-allocator-service"
readme = "README.md"
requires-python = ">=3.9"
version = "0.0.3a2"

[project.optional-dependencies]
dev = ["twine", "build", "pytest", "ruff", "pytest-cov"]

[project.scripts]
lablink-allocator = "lablink_allocator_service.main:main"
generate-init-sql = "lablink_allocator_service.generate_init_sql:main"
lablink-validate-config = "lablink_allocator_service.validate_config:main"

[tool.setuptools.package-data]
lablink_allocator_service = ["terraform/**/*", "conf/*.yaml"]

[project.urls]
Homepage = "https://github.com/talmolab/lablink"
Issues = "https://github.com/talmolab/lablink/issues"

[tool.ruff.lint]
ignore = [
  "E721", # Use isinstance() for type comparisons
  "E731", # Do not assign a lambda expression
  "E741", # Ambiguous variable name
  "F405", # May be undefined, or defined from star imports
]
select = ["E", "F", "W"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "E402"]
"tests/*" = ["F403", "F405", "E501"]
